requirements.txt
tensorflow-gpu==2.10
numpy
pandas
opencv
pip
Pillow==9.5.0
arabic-reshaper
python-bidi
